+++ app/src/main/java/com/TellMeUp/tellmeapp/MainActivity.kt
@@ -1,7 +1,7 @@
 /**
  * @file: MainActivity.kt
  * @description: Main entry point of the application with Compose UI
- * @dependencies: Hilt, Compose Navigation
+ * @dependencies: Hilt, TellMeAppTheme, MainScreen
  * @created: 2026-05-08
  */
 
@@ -11,13 +11,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
+import com.TellMeUp.tellmeapp.ui.screen.main.MainScreen
 import com.TellMeUp.tellmeapp.ui.theme.TellMeAppTheme
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -27,30 +21,9 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            TellMeAppTheme {
-                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    Greeting(
-                        name = "TellMeApp",
-                        modifier = Modifier.padding(innerPadding)
-                    )
-                }
+            TellMeAppTheme(darkTheme = true) {
+                MainScreen()
             }
         }
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    TellMeAppTheme {
-        Greeting("TellMeApp")
     }
 }